Fix the TypeScript bug I was asked to fix :)
Before, the type of the `useState` was inferred to be `{ start: null, end: null }`.
That's why TypeScript errored later on: You tried to call `.split` on `null`, which is nonsense.

Now in reality, you type is more like `{ start: string | null, end: string | null }`.
TypeScript has no chance of inferring that, but once you specify it (`useState` takes it as a type parameter),
everything works fine :D